Add lib/chainer/device.rb and move Device classes to there
